requirements.txt
gym==0.17.2
matplotlib==3.3.0
numpy==1.19.0
pandas==1.0.5
pillow==7.2.0
scipy==1.5.1
seaborn==0.10.1
tensorboard==2.3.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu